refactor: :truck: Cambiar nombre de carpeta log a logs
