+++ crates/hardware/src/logs/mod.rs
